move shared/middleware to dashboard/lib
