+++ bin/cron/applab_datasets/spotify
@@ -11,7 +11,7 @@ abort 'Script already running' unless only_one_running?(__FILE__)
 
 require_relative '../../../deployment'
 require 'ostruct'
-require_relative '../../../shared/middleware/helpers/firebase_helper'
+require_relative '../../../dashboard/lib/middleware/helpers/firebase_helper'
 require 'cdo/profanity_filter'
 require 'rest-client'
