+++ dashboard/config/application.rb
@@ -5,13 +5,13 @@ require 'rails/all'
 require 'cdo/geocoder'
 require 'cdo/properties'
 require 'varnish_environment'
-require_relative '../../shared/middleware/files_api'
-require_relative '../../shared/middleware/channels_api'
-require_relative '../../shared/middleware/tables_api'
+require_relative '../lib/middleware/files_api'
+require_relative '../lib/middleware/channels_api'
+require_relative '../lib/middleware/tables_api'
 require 'shared_resources'
-require_relative '../../shared/middleware/net_sim_api'
-require_relative '../../shared/middleware/sound_library_api'
-require_relative '../../shared/middleware/animation_library_api'
+require_relative '../lib/middleware/net_sim_api'
+require_relative '../lib/middleware/sound_library_api'
+require_relative '../lib/middleware/animation_library_api'
 
 require 'bootstrap-sass'
 require 'cdo/hash'
