+++ dashboard/lib/middleware/helpers/animation_bucket.rb
