+++ dashboard/lib/middleware/helpers/asset_bucket.rb
