+++ dashboard/lib/middleware/helpers/file_bucket.rb
