+++ dashboard/lib/middleware/helpers/firebase_helper.rb
