+++ dashboard/lib/middleware/helpers/library_bucket.rb
