+++ dashboard/lib/middleware/helpers/profanity_privacy_helper.rb
@@ -1,6 +1,6 @@
 require 'cdo/share_filtering'
-require_relative './bucket_helper.rb'
-require_relative './source_bucket.rb'
+require_relative './bucket_helper'
+require_relative './source_bucket'
 require 'cdo/firehose'
 
 class ProfanityPrivacyError < StandardError
