+++ dashboard/lib/middleware/helpers/pusher_api.rb
