+++ dashboard/lib/middleware/helpers/redis_property_bag.rb
