+++ dashboard/lib/middleware/helpers/redis_table.rb
