+++ dashboard/lib/middleware/helpers/sharded_redis_factory.rb
