+++ dashboard/lib/middleware/helpers/table.rb
