+++ dashboard/lib/middleware/sound_library_api.rb
