+++ dashboard/lib/middleware/tables_api.rb
@@ -7,12 +7,15 @@ class TablesApi < Sinatra::Base
   set :mustermann_opts, check_anchors: false
 
   helpers do
+    load(CDO.dir('shared', 'middleware', 'helpers', 'core.rb'))
+  end
+
+  helpers do
     [
-      'core.rb',
       'table.rb',
       'firebase_helper.rb',
     ].each do |file|
-      load(CDO.dir('shared', 'middleware', 'helpers', file))
+      load(CDO.dir('dashboard', 'lib', 'middleware', 'helpers', file))
     end
   end
